Added intercepted pages ...
